+++ cx-portal/src/features/serviceMarketplace/serviceApiSlice.ts
@@ -74,12 +74,26 @@ export type SubscriptionServiceRequest = {
   body: SubscriptionRequestBody[]
 }
 
+export type ServiceBody = {
+  page: number
+  serviceType: string
+  sortingType: string
+}
+
 export const apiSlice = createApi({
   reducerPath: 'rtk/apps/service',
   baseQuery: fetchBaseQuery(apiBaseQuery()),
   endpoints: (builder) => ({
-    fetchServices: builder.query<ServiceRequestAPIResponse, number>({
-      query: (page) => `/api/services/active?size=${PAGE_SIZE}&page=${page}`,
+    fetchServices: builder.query<ServiceRequestAPIResponse, ServiceBody>({
+      query: (body) => {
+        const serviceType = `serviceTypeId=${body.serviceType}`
+        const sortingType = `sorting=${body.sortingType}`
+        return {
+          url: `/api/services/active?size${PAGE_SIZE}&page=${body.page}&${
+            body.serviceType && serviceType
+          }&${body.sortingType && sortingType}`,
+        }
+      },
     }),
     fetchService: builder.query<ServiceRequest, string>({
       query: (serviceId) =>